refactor(chat): mirror bisq2's ChatMessage and ChatChannel roots and retype the shared chat composables to them
